Flatten form/config/ into form/ to match cell/ conventions
