+++ javascript/packages/core/components/form/types.ts
@@ -1,4 +1,6 @@
 import type { FORM_ERROR } from 'final-form';
+import type { ComponentType } from 'react';
+import type { BaseFieldProps } from '#core/components/form/fields/types';
 import type { DeepPartial } from '#core/types/utility-types';
 
 export type FormData = Record<string, unknown>;
@@ -22,31 +24,6 @@ export interface FormProps<FieldValues extends FormData = FormData> {
    * Optional render prop for wrapping the form element.
    * When provided, the form element is passed to this function, allowing
    * components outside the form element to access form state via useFormState.
-   *
-   * @example
-   * ```tsx
-   * // Form with external submit button in wrapper
-   * <Form
-   *   id="my-form"
-   *   onSubmit={handleSubmit}
-   *   render={(formElement) => (
-   *     <div>
-   *       {formElement}
-   *       <footer>
-   *         <button type="submit" form="my-form">Submit</button>
-   *       </footer>
-   *     </div>
-   *   )}
-   * >
-   *   <StringField name="email" label="Email" />
-   * </Form>
-   *
-   * // Standalone form (no render prop needed)
-   * <Form onSubmit={handleSubmit}>
-   *   <StringField name="email" label="Email" />
-   *   <button type="submit">Submit</button>
-   * </Form>
-   * ```
    */
   render?: (formElement: React.ReactNode) => React.ReactNode;
 
@@ -55,15 +32,6 @@ export interface FormProps<FieldValues extends FormData = FormData> {
    *
    * @note `right` is usually reserved for form actions (e.g., submit button).
    * @note `left` is usually reserved for secondary info, status text.
-   *
-   * @example
-   * ```tsx
-   * // Object with left and right content
-   * <Form footer={{ right: <SubmitButton>Save</SubmitButton>, left: <span>Last saved 2m ago</span> }}>
-   *
-   * // ReactNode for full control
-   * <Form footer={<MyCustomFooter />}>
-   * ```
    */
   footer?: { left?: React.ReactNode; right?: React.ReactNode } | React.ReactNode;
 }
@@ -123,3 +91,114 @@ export interface ArrayFieldOptions {
   minItems?: number;
   readOnly?: boolean;
 }
+
+export enum FieldType {
+  STRING = 'string',
+  NUMBER = 'number',
+  BOOLEAN = 'boolean',
+  SELECT = 'select',
+  CHECKBOX = 'checkbox',
+  RADIO = 'radio',
+  DATE = 'date',
+  TEXTAREA = 'textarea',
+  URL = 'url',
+  MAP = 'map',
+  MARKDOWN = 'markdown',
+}
+
+/**
+ * Module-augmentation extension point for consumer field configs.
+ *
+ * @example
+ * ```ts
+ * declare module '@michelangelo-ai/core' {
+ *   interface FieldConfigExtensions {
+ *     'select-hive': { type: 'select-hive'; cluster: string };
+ *   }
+ * }
+ * ```
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type
+export interface FieldConfigExtensions {}
+
+export type SharedFieldConfig = {
+  label?: string;
+  required?: boolean;
+  disabled?: boolean;
+  readOnly?: boolean;
+  placeholder?: string;
+  description?: string;
+  caption?: string;
+  defaultValue?: unknown;
+  initialValue?: unknown;
+};
+
+type StringFieldConfig = SharedFieldConfig & {
+  type: FieldType.STRING | 'string';
+  multi?: boolean;
+};
+
+export type BuiltinFieldConfig = StringFieldConfig;
+
+export type FieldConfig = BuiltinFieldConfig | (SharedFieldConfig & { type: string });
+
+export type FieldRendererProps = BaseFieldProps & {
+  config: Record<string, unknown>;
+};
+
+export type FieldRenderer = ComponentType<FieldRendererProps>;
+
+/**
+ * Module-augmentation extension point for consumer layout types.
+ *
+ * @example
+ * ```ts
+ * declare module '@michelangelo-ai/core' {
+ *   interface LayoutConfigExtensions {
+ *     tabs: { type: 'tabs'; items: LayoutItem[] };
+ *   }
+ * }
+ * ```
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type
+export interface LayoutConfigExtensions {}
+
+type GroupLayoutConfig = {
+  type: 'group';
+  label?: string;
+  description?: string;
+  tooltip?: string;
+  collapsible?: boolean;
+  items: LayoutItem[];
+};
+
+type RowLayoutConfig = {
+  type: 'row';
+  name?: string;
+  span?: number[];
+  items: LayoutItem[];
+};
+
+type GridLayoutConfig = {
+  type: 'grid';
+  items: LayoutItem[];
+};
+
+export type BuiltinLayoutConfig = GroupLayoutConfig | RowLayoutConfig | GridLayoutConfig;
+
+export type LayoutConfig = BuiltinLayoutConfig | { type: string; items: LayoutItem[] };
+
+/** A layout item is either a layout config object or a bare field path string. */
+export type LayoutItem = LayoutConfig | string;
+
+export type LayoutRendererProps = {
+  config: LayoutConfig;
+  renderItems: (items: LayoutItem[]) => React.ReactNode;
+};
+
+export type LayoutRenderer = ComponentType<LayoutRendererProps>;
+
+export type FormConfig = {
+  entities: Record<string, FieldConfig>;
+  layout: LayoutItem[];
+};
