+++ javascript/packages/core/components/views/sandbox/sandbox.tsx
@@ -1,10 +1,10 @@
 import { Block } from 'baseui/block';
 import { HeadingXXLarge } from 'baseui/typography';
 
-import { ConfigDrivenForm } from '#core/components/form/config/config-driven-form';
+import { ConfigDrivenForm } from '#core/components/form/config-driven-form';
 import { MainViewContainer } from '#core/components/views/main-view-container';
 
-import type { FormConfig } from '#core/components/form/config/types';
+import type { FormConfig } from '#core/components/form/types';
 
 const SAMPLE_FORM_CONFIG: FormConfig = {
   entities: {
